separate content-scripts and components
